{{cookiecutter.project_slug}}/requirements.txt
coverage
Django
django-cors-headers
django-debug-toolbar
django-filter
django-imagekit
django-model-utils
django-rest-auth
django-storages
django-extensions
djangorestframework
Pillow
psycopg2-binary
pytest
pytest-asyncio
pytest-cov
pytest-django
pytest-factoryboy
pytest-mock
factory-boy
django-factory-generator